完成了 临时 U  V 的计算

+++ matrix/matrix/setMatrix.h
@@ -51,7 +51,7 @@ void setSamplingPointDensity(float *density, int numOfPoints){
 		return;
 	for (int i = 0; i < numOfPoints; i++){
 
-		*(density + i) = 1 / numOfPoints;
+		*(density + i) = (float) 1 / numOfPoints;
 	}
 	return;
 }
